Improved data presantation-play button update
when 'play' button is pressed, instead of showing real figures,
displayed is presantation of images- it shows overall ozone migration,
and is lighter to display => faster presantation, less computation
power.

Update to addColorbar(), update to generateMap() so that it's compatible
with new presantations method.

Todo: write unit test and comments.